+++ CHANGELOG.md
@@ -14,13 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Commands to switch between tabs ([#5](https://github.com/benji300/joplin-note-tabs/issues/5))
 - Option to disable to-do checkboxes on tabs ([#6](https://github.com/benji300/joplin-note-tabs/issues/6))
 - Possibility to pin note from editor context menu ([#9](https://github.com/benji300/joplin-note-tabs/pull/9) by [@ambrt](https://github.com/ambrt))
-- Command to pin note(s) to tabs from note list context menu ([#12](https://github.com/benji300/joplin-note-tabs/pull/12)
+- Command to pin note(s) from note list context menu ([#12](https://github.com/benji300/joplin-note-tabs/pull/12)
 
 ### Changed
 
 - Decreased jpl size
 - Added description to place tabs to README ([#4](https://github.com/benji300/joplin-note-tabs/pull/4) by [@amandamcg](https://github.com/amandamcg))
 - Allow moving unpinned tabs
+- Default value of advanced style settings is now `default` for all
 
 ### Fixed